Start of new schedule and a few new icons released under creative commons as svg but converted to png files.
The new schedule interface is designed so that you specify the days of the week you want the campaign to run as well as the times you
want it to run.

This creates an entry in the schedule table (changes to login.php to create the table if it doesn't exist).

A cronjob is added which runs every minute and checks if there are any campaigns it should be starting or stopping.

I haven't added the cronjob yet.