Make it so that you can have multiple host names.
Basically the way it works is that the first time you load it, it checks if the table exists.

If it doesn't it creates it and populates it with all of the details from the config files.

If so it uses it.

I have also completed saving to the database for config changes.

The only information that is now written to the config file is the main SmoothTorque MySQL host/user/pass.

That needs to stay in the file because we can't read the config information without knowing where to read it from.

I still need to create an interface to create new hosts (i.e. copy from x).

The current configs are copied into a url of default - if there is no entry for the current server then it will use default.